Added the VariableCategorizerReducer to DooSelection.
This is the first version of this reducer. Please test extensive!

This is a Reducer derived from Reducer. It writes a new variable categorizing a given variable into N bins with an equal number of entries. To do so, it first calculates the N-quantiles for the given variable in the provided variable range. Then it sorts every event into one of the quantiles.